Unify main.js and debug.js in Code/Extension
