+++ Documentation/2. Developer guide/4. Features.md
@@ -19,6 +19,7 @@
 - Create a language server
 - Connect it to language client using stdin stdio
 - Connect it to language client using socket 5007
+- Reconnect it to language client using socket 5007
 
 ### Server.java
